Share Vite logger logic between web and desktop
