+++ vite.config.ts
@@ -6,25 +6,15 @@ import react from '@vitejs/plugin-react'
 import version from 'vite-plugin-package-version'
 import topLevelAwait from 'vite-plugin-top-level-await'
 import viteTsconfigPaths from 'vite-tsconfig-paths'
-import { createLogger } from 'vite'
 import { configDefaults, defineConfig } from 'vitest/config'
-import { indexHtmlCsp } from './vite.base.config'
+import { createCustomLogger, indexHtmlCsp } from './vite.base.config'
 import { nodePolyfills } from 'vite-plugin-node-polyfills'
-
-const publicAssetWarning =
-  'Assets in public directory cannot be imported from JavaScript'
-const logger = createLogger()
-const originalWarn = logger.warn.bind(logger)
-logger.warn = (msg, opts) => {
-  if (msg.includes(publicAssetWarning)) return
-  originalWarn(msg, opts)
-}
 
 export default defineConfig(({ command, mode }) => {
   const runMillion = process.env.RUN_MILLION
 
   return {
-    customLogger: logger,
+    customLogger: createCustomLogger(),
     define: {
       'import.meta.env.VERCEL_ENV': JSON.stringify(process.env.VERCEL_ENV),
     },
